Deprecate torch.cross default behaviour
Long overdue this one. We may be able to change it in a few years :hopeful:.

**BC-breaking note**

This PR deprecates `torch.cross`'s default dim in favor of
`torch.linalg.cross`.
A upgrade guide is added to the documentation for `torch.cross`.

Note this PR DOES NOT remove `torch.cross`.

Fixes https://github.com/pytorch/pytorch/issues/108664

ghstack-source-id: fb21a77f924bea8669543611b3a98ad2e5f2c117
Pull Request resolved: https://github.com/pytorch/pytorch/pull/108760